Implement bus in terms of futures
This mostly works, with the exception of a race where writers may on
occasion race with readers, and not wake them up even though a write has
happened. This causes a deadlock because the writers can't write more
until the readers have run, and the readers are parked, and haven't been
told to resume reading. In the non-future implementation, this works out
because readers will periodically wake up (using park_timeout), but this
is not the case for the Futures version, which doesn't have such a
timeout.