T must be Sync; fixes #13
@stjepang points out in #13 that the cloning of T that happens in each
`BufReader` can happen concurrently in different threads, and T must
therefore be Sync. It *not* being Sync leads to undefined behavior,
which is a bug. I am *not* bumping the major version number, as this is
in fact a bugfix, even though it may break existing code (which would
then also currently be buggy).