Remove broken async impl. Fixes #14
This feature was broken (as pointed out in #14), and requires a decent
amount of work to fix. We already track thread handles to receivers that
are blocking, but they aren't *guaranteed* to be woken up (which is why
they use park_timeout). We'd also have to change the code to take *both*
thread and task handles, which'd take some work. It's not undoable, but
I'll let this go for now.