Implement futures-compatible async bus.
The implementation is pretty crappy! It's full of TODO comments. I
didn't reorganize the code in a way that preserves much of the original
semantic clarity, which was definitely the first mistake. It works, and
does demonstrate one way of factoring around sometimes blocking and
sometimes not.

The tests might not be comprehensive enough either. I'm not sure.